+++ pkgs/tools/nix/nixos-render-docs/src/nixos_render_docs/docbook.py
@@ -1,0 +1,12 @@
+_xml_id_translate_table = {
+    ord('*'): ord('_'),
+    ord('<'): ord('_'),
+    ord(' '): ord('_'),
+    ord('>'): ord('_'),
+    ord('['): ord('_'),
+    ord(']'): ord('_'),
+    ord(':'): ord('_'),
+    ord('"'): ord('_'),
+}
+def make_xml_id(s: str) -> str:
+    return s.translate(_xml_id_translate_table)
